refactor: ♻️ Move dto and resolver to interface-adapters
