+++ README.md
@@ -54,14 +54,15 @@ You can use `npx nx list` to get a list of installed plugins. Then, run `npx nx 
 
 ## Architecture
 
-| Layer                    | Description                                                                                                                                                                                                                                                                                                                                                               |
-| ------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| dto(presentation)        | Define DTOs for GraphQL schema.                                                                                                                                                                                                                                                                                                                                           |
-| resolver(presentation)   | Define GraphQL schema and resolver.                                                                                                                                                                                                                                                                                                                                       |
-| mongoose(infrastructure) | Implements the repository interfaces defined in the domain layer using Mongoose as the ODM (Object Document Mapper). <br/>Includes Mongoose Schema definitions, database connection management, and concrete implementations of repository interfaces (e.g., MongooseUserRepository).                                                                                     |
-| use-case(application)    | Define business use cases and encapsulate business logic.                                                                                                                                                                                                                                                                                                                 |
-| entity(domain)           | Define core business entities and business rules.<br/> Maintain entity independence from database and framework.                                                                                                                                                                                                                                                          |
-| repository(domain)       | Interfaces (or abstract classes), which define methods for manipulating data without concern for specific database implementations. <br/>By defining this interface, we can decouple database access: the specific details of data access will be done by implementation classes, such as specific implementations using tools like Mongoose, TypeORM, Prisma, and so on. |
+| Layer                        | Description                                                                                                                                                                                                                                                                                                                                                               |
+| ---------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| resolver(interface-adapters) | Define GraphQL schema and resolver.                                                                                                                                                                                                                                                                                                                                       |
+| dto(interface-adapters)      | Define DTOs for GraphQL schema.                                                                                                                                                                                                                                                                                                                                           |
+| mongoose(infrastructure)     | Implements the repository interfaces defined in the domain layer using Mongoose as the ODM (Object Document Mapper). <br/>Includes Mongoose Schema definitions, database connection management, and concrete implementations of repository interfaces (e.g., MongooseUserRepository).                                                                                     |
+| service(application)         | As the core of the application layer, it mainly interacts with the domain layer and interface-adapter layer.<br/> If you migrate to a non-NestJS architecture in the future (e.g. other frameworks or microservices), the application tier code can be left unaffected.                                                                                                   |
+| use-case(application)        | Define business use cases and encapsulate business logic.                                                                                                                                                                                                                                                                                                                 |
+| entity(domain)               | Define core business entities and business rules.<br/> Maintain entity independence from database and framework.                                                                                                                                                                                                                                                          |
+| repository(domain)           | Interfaces (or abstract classes), which define methods for manipulating data without concern for specific database implementations. <br/>By defining this interface, we can decouple database access: the specific details of data access will be done by implementation classes, such as specific implementations using tools like Mongoose, TypeORM, Prisma, and so on. |
 
 ## Useful links
