+++ libs/users/interface-adapters/README.md
@@ -1,0 +1,7 @@
+# users-interface-adapters
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test users-interface-adapters` to execute the unit tests via [Jest](https://jestjs.io).
