+++ libs/users/interface-adapters/eslint.config.js
@@ -1,0 +1,3 @@
+const baseConfig = require('../../../eslint.config.js');
+
+module.exports = [...baseConfig];
