+++ libs/users/interface-adapters/jest.config.ts
@@ -1,10 +1,10 @@
 export default {
-  displayName: 'users-presentation-resolver',
-  preset: '../../../../jest.preset.js',
+  displayName: 'users-interface-adapters',
+  preset: '../../../jest.preset.js',
   testEnvironment: 'node',
   transform: {
     '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../../coverage/libs/users/presentation/resolver',
+  coverageDirectory: '../../../coverage/libs/users/interface-adapters',
 };
