+++ libs/users/interface-adapters/src/index.ts
@@ -1,0 +1,4 @@
+// dto
+export * from './lib/dto/user.dto';
+// resolver
+export * from './lib/resolver/users.resolver';
