--- libs/users/presentation/dto/README.md
@@ -1,7 +1,0 @@
-# users-presentation-dto
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test users-presentation-dto` to execute the unit tests via [Jest](https://jestjs.io).
