--- libs/users/presentation/dto/jest.config.ts
@@ -1,10 +1,0 @@
-export default {
-  displayName: 'users-presentation-dto',
-  preset: '../../../../jest.preset.js',
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../../coverage/libs/users/presentation/dto',
-};
