--- libs/users/presentation/dto/project.json
@@ -1,9 +1,0 @@
-{
-  "name": "users-presentation-dto",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/users/presentation/dto/src",
-  "projectType": "library",
-  "tags": [],
-  "// targets": "to see all targets run: nx show project users-presentation-dto --web",
-  "targets": {}
-}
