--- libs/users/presentation/resolver/README.md
@@ -1,7 +1,0 @@
-# users-presentation-resolver
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test users-presentation-resolver` to execute the unit tests via [Jest](https://jestjs.io).
