--- libs/users/presentation/resolver/eslint.config.js
@@ -1,3 +1,0 @@
-const baseConfig = require('../../../../eslint.config.js');
-
-module.exports = [...baseConfig];
