--- libs/users/presentation/resolver/project.json
@@ -1,9 +1,0 @@
-{
-  "name": "users-presentation-resolver",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/users/presentation/resolver/src",
-  "projectType": "library",
-  "tags": [],
-  "// targets": "to see all targets run: nx show project users-presentation-resolver --web",
-  "targets": {}
-}
