--- libs/users/presentation/resolver/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/users.resolver';
