+++ tsconfig.base.json
@@ -20,15 +20,14 @@
       "@shared/infrastructure-mongoose": [
         "libs/shared/infrastructure/mongoose/src/index.ts"
       ],
-      "@users/domain": ["libs/users/domain/src/index.ts"],
       "@users/application": ["libs/users/application/src/index.ts"],
+      "@users/domain": ["libs/users/domain/src/index.ts"],
       "@users/infrastructure-mongoose": [
         "libs/users/infrastructure/mongoose/src/index.ts"
       ],
-      "@users/presentation-dto": ["libs/users/presentation/dto/src/index.ts"],
-      "@users/presentation-resolver": [
-        "libs/users/presentation/resolver/src/index.ts"
-      ]
+      "@users/interface-adapters": [
+        "libs/users/interface-adapters/src/index.ts"
+      ],
     }
   },
   "exclude": ["node_modules", "tmp"]
